# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clinic/whatsapp/consent:
    get:
      operationId: getWhatsappConsent
      tags: [clinic]
      summary: Read WhatsApp consent for a patient
      responses:
        "200":
          description: Consent state
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
    post:
      operationId: updateWhatsappConsent
      tags: [clinic]
      summary: Opt in or out of clinic WhatsApp messages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WhatsappConsentInput"
      responses:
        "200":
          description: Consent saved
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /clinic/whatsapp/outbox:
    get:
      operationId: listWhatsappOutbox
      tags: [clinic]
      summary: List WhatsApp delivery history
      responses:
        "200":
          description: Outbox messages
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
    post:
      operationId: enqueueWhatsappMessage
      tags: [clinic]
      summary: Queue a consented WhatsApp message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WhatsappOutboxInput"
      responses:
        "201":
          description: Queued message
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...
components:
...
  schemas:
...
    WhatsappConsentInput:
      type: object
      required: [optedIn]
      properties:
        clientId: { type: string }
        phone: { type: string }
        optedIn: { type: boolean }
    WhatsappOutboxInput:
      type: object
      required: [clientId, eventType, idempotencyKey, fallbackText]
      properties:
        clientId: { type: string }
        eventType: { type: string }
        idempotencyKey: { type: string }
        payload: { type: object, additionalProperties: true }
        fallbackText: { type: string }
...