# lib/api-zod/src/generated/types/whatsappOutboxInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WhatsappOutboxInputPayload } from './whatsappOutboxInputPayload';

export interface WhatsappOutboxInput {
  clientId: string;
  eventType: string;
  idempotencyKey: string;
  payload?: WhatsappOutboxInputPayload;
  fallbackText: string;
}